pyproject.toml
[project]
name = "hackathon-llamaindex-mvp"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "llama-index",
    "python-dotenv",
]